refactoring and renaming to increase clarity
