+++ security-monitor/src/core/architecture/mod.rs
@@ -2,8 +2,8 @@
 // SPDX-FileContributor: Wojciech Ozga <woz@zurich.ibm.com>, IBM Research - Zurich
 // SPDX-License-Identifier: Apache-2.0
 pub use riscv::{
-    AceExtension, BaseExtension, FpRegisters, GpRegister, GpRegisters, HartState, HsmExtension, IpiExtension,
-    RfenceExtension, SbiExtension, SrstExtension, TrapReason,
+    decode_result_register, AceExtension, BaseExtension, FpRegisters, GpRegister, GpRegisters, HartState, HsmExtension,
+    IpiExtension, RfenceExtension, SbiExtension, SrstExtension, TrapReason,
 };
 
 mod riscv;
