+++ security-monitor/src/core/architecture/riscv/mod.rs
@@ -1,12 +1,14 @@
 // SPDX-FileCopyrightText: 2023 IBM Corporation
 // SPDX-FileContributor: Wojciech Ozga <woz@zurich.ibm.com>, IBM Research - Zurich
 // SPDX-License-Identifier: Apache-2.0
+pub use compressed_instructions::decode_result_register;
 pub use fp_registers::FpRegisters;
 pub use gp_registers::{GpRegister, GpRegisters};
 pub use hart_state::HartState;
 pub use sbi::{AceExtension, BaseExtension, HsmExtension, IpiExtension, RfenceExtension, SbiExtension, SrstExtension};
 pub use trap_reason::TrapReason;
 
+mod compressed_instructions;
 mod fp_registers;
 mod gp_registers;
 mod hart_state;
