+++ security-monitor/src/core/architecture/riscv/trap_reason.rs
@@ -1,8 +1,8 @@
 // SPDX-FileCopyrightText: 2023 IBM Corporation
 // SPDX-FileContributor: Wojciech Ozga <woz@zurich.ibm.com>, IBM Research - Zurich
 // SPDX-License-Identifier: Apache-2.0
-use crate::core::arch::riscv::sbi::SbiExtension;
-use crate::core::arch::HartState;
+use crate::core::architecture::riscv::sbi::SbiExtension;
+use crate::core::architecture::HartState;
 
 #[derive(Debug)]
 pub enum TrapReason {
