+++ security-monitor/src/core/memory_protector/mmu/paging_system.rs
@@ -1,8 +1,8 @@
 // SPDX-FileCopyrightText: 2023 IBM Corporation
 // SPDX-FileContributor: Wojciech Ozga <woz@zurich.ibm.com>, IBM Research - Zurich
 // SPDX-License-Identifier: Apache-2.0
+use crate::core::memory_layout::ConfidentialVmVirtualAddress;
 use crate::core::memory_protector::PageSize;
-use crate::core::transformations::ConfidentialVmVirtualAddress;
 use riscv::register::hgatp::HgatpMode;
 
 // TODO: add more 2nd-level paging systems corresponding to 3 and 4 level page
