+++ security-monitor/src/core/mod.rs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2023 IBM Corporation
 // SPDX-FileContributor: Wojciech Ozga <woz@zurich.ibm.com>, IBM Research - Zurich
 // SPDX-License-Identifier: Apache-2.0
-pub mod arch;
+pub mod architecture;
 pub mod control_data;
 pub mod memory_layout;
 pub mod memory_protector;
