+++ security-monitor/src/core/transformations/guest_load_page_fault_result.rs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2023 IBM Corporation
 // SPDX-FileContributor: Wojciech Ozga <woz@zurich.ibm.com>, IBM Research - Zurich
 // SPDX-License-Identifier: Apache-2.0
-use crate::core::arch::{GpRegister, HartState};
+use crate::core::architecture::{GpRegister, HartState};
 use crate::core::transformations::GuestLoadPageFaultRequest;
 
 pub struct GuestLoadPageFaultResult {
