+++ security-monitor/src/core/transformations/mmio_store_request.rs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2023 IBM Corporation
 // SPDX-FileContributor: Wojciech Ozga <woz@zurich.ibm.com>, IBM Research - Zurich
 // SPDX-License-Identifier: Apache-2.0
-use crate::core::arch::GpRegister;
+use crate::core::architecture::GpRegister;
 
 pub struct MmioStoreRequest {
     code: usize,
