+++ security-monitor/src/core/transformations/opensbi_request.rs
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2023 IBM Corporation
 // SPDX-FileContributor: Wojciech Ozga <woz@zurich.ibm.com>, IBM Research - Zurich
 // SPDX-License-Identifier: Apache-2.0
-use crate::core::arch::{GpRegister, HartState};
+use crate::core::architecture::{GpRegister, HartState};
 
 pub struct OpensbiRequest {
     pub regs: opensbi_sys::sbi_trap_regs,
